Re-use of_age method and put each class in seperate file
