+++ person.rb
@@ -9,7 +9,7 @@ class Person
   end
 
   def can_use_services?
-    @age >= 18 || @parent_permission == true
+    of_age? || @parent_permission == true
   end
 
   def say_age
@@ -23,24 +23,3 @@ class Person
   end
 end
 
-class Student < Person
-  def initialize(classroom, age, _name: 'Unknown', _parent_permission: true)
-    super(age, _name: 'Unknown', _parent_permission: true)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    "¯\(ツ)/¯"
-  end
-end
-
-class Teacher < Person
-  def initialize(specialization, age, _name: 'Unknown', _parent_permission: true)
-    super(age, _name: 'Unknown', _parent_permission: true)
-    @specialization = specialization
-  end
-
-  def can_use_services?
-    true
-  end
-end
