+++ student.rb
@@ -1,0 +1,14 @@
+require './person.rb'
+
+class Student < Person
+    def initialize(classroom, age, _name: 'Unknown', _parent_permission: true)
+      super(age, _name: 'Unknown', _parent_permission: true)
+      @classroom = classroom
+    end
+  
+    def play_hooky
+      "¯\(ツ)/¯"
+    end
+  end
+student = Student.new("Hey", 10)
+puts student.can_use_services?
